Let `require("b")` resolve to b/index.js
This commit changes b's "exports" config to be more in line
whith what usually happens for ESM-only: there isn't an explicit
condition that filter "import", but it just points to the ESM file
regardless of how it's being loaded.

This is necessary to show how the new hazard case comes
up in Node.js, that previously was an error. As shown by the
webpack test, the hazard was already present when using bundlers.

Results for Node.js without the flag:
- index.mjs: [ 'PASS', 'HAZARD' ]
- index.cjs: [ 'ERR_REQUIRE_ESM', 'PASS' ]

Results for Node.js with the flag:
- index.mjs: [ 'PASS', 'HAZARD' ]
- index.cjs: [ 'HAZARD', 'PASS' ]

Results for Webpack:
- index.mjs: [ 'PASS', 'HAZARD' ]
- index.cjs: [ 'HAZARD', 'PASS' ]